Renamed class to UseAutoDispose
